log: cast argument to double

+++ utils/log.h
@@ -1,2 +1,2 @@
 #include <stdio.h>
-#define LOG(what) fprintf(stderr, "%s %s = %f\n", __FUNCTION__, #what, what);
+#define LOG(what) fprintf(stderr, "%s %s = %f\n", __FUNCTION__, #what, (double)(what));
